refactor: 패키지 분리
- domain 내 클래스들 card, gamer, rule로 분류
