+++ README.md
@@ -50,3 +50,29 @@
 - [x] 합계가 21을 초과하지 않고, 21에 더 가까운 사람이 승리한다.
 
 - [x] 카드의 합계가 같을 경우 무승부다.
+
+# 2단계 추가 요구사항
+
+- [ ] 플레이어는 게임을 시작할 때 배팅 금액을 정해야 한다.
+    - [ ] 베팅 금액이 0원 이하라면 예외처리
+- [ ] 플레이어의 처음 두 장의 카드 합이 21일 경우 블랙잭
+- [ ] 플레이어가 블랙잭일 경우 베팅 금액의 1.5 배를 딜러에게 받는다.
+    - [ ] 딜러도 블랙잭일 경우에는 베팅한 금액을 돌려받는다.
+- [ ] 딜러가 21을 초과하면 플레이어들은 승리해 베팅 금액을 받는다.
+    - [ ] 플레이어도 21을 초과한 경우에는 베팅 금액을 잃는다.
+
+## 플레이어의 베팅금액이 10000원일 때의 경우
+
+| 딜러    | 플레이어  | 딜러 수익  | 플레이어 수익 |
+|-------|-------|--------|---------|
+| 블랙잭   | 블랙잭   | 0      | 0       |
+| 블랙잭   | 일반 점수 | 10000  | -10000  |
+| 블랙잭   | 버스트   | 10000  | -10000  |
+|       |       |        |         |
+| 일반 점수 | 블랙잭   | -15000 | 15000   |
+| 일반 점수 | 일반 점수 | ??     | ??      |
+| 일반 점수 | 버스트   | 10000  | -10000  |
+|       |       |        |         |
+| 버스트   | 블랙잭   | -15000 | 15000   |
+| 버스트   | 일반 점수 | -10000 | 10000   |
+| 버스트   | 버스트   | 10000  | -10000  |
