+++ src/main/java/domain/card/CardsInitializer.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import java.util.List;
