+++ src/main/java/domain/card/StandardCardsInitializer.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import java.util.ArrayList;
 import java.util.Arrays;
