+++ src/main/java/domain/gamer/Player.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.gamer;
 
 public class Player extends Gamer {
