+++ src/main/java/domain/rule/BlackjackMatchResult.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.rule;
 
 public enum BlackjackMatchResult {
     WIN("승"),
