+++ src/test/java/domain/gamer/PlayerTest.java
@@ -1,9 +1,13 @@
-package domain;
+package domain.gamer;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import domain.card.Card;
+import domain.card.CardDeck;
+import domain.card.CardRank;
+import domain.card.CardSymbol;
 import fixture.CardsInitializerFixture;
 import java.util.ArrayList;
 import java.util.List;
