+++ src/test/java/fixture/HandFixture.java
@@ -1,7 +1,7 @@
 package fixture;
 
-import domain.Card;
-import domain.Hand;
+import domain.card.Card;
+import domain.card.Hand;
 
 public class HandFixture {
     public static Hand createHand(Card... cards) {
